Fix casting problem if first cast is not the right one for tile width
In Kotlin, use Float::class.java and use it as 2nd parameter of
TiledMap.properties.get("tilewidth", ...).

This way, we didn't have to cast it to Integer first then use .toFloat()
later. The point is that we really have to cast to the right (the one
that used by Tiled application) one otherwise there will be error.