Refactor screen size updatable class. Fix updating lag.
Screen size changesd is now refactored into interface class. Still the
limitation is that it cannot straight away starts as full screen, it
needs to be activated by user manually later. This is in order to let
the code detect and update screen size change as such.

Updating now carried out as soon as needed according to the default
frame rate set on libgdx (60fps).