Add checking when collide with tile in front
This created another sensor in front but smaller to check if front part
of player collides with tile. If collided then apply force (with spin,
but due to the core code doesn't use Sprite class so we don't have
ability to rotate the sprite, anyway we live with this for now).

When front collided, then stop via ScreenStopper.