Add navigating selection support in level selection screen
Navigating selection support include all possible input; keyboard,
mouse, and controller with 4 directional wrapped-around.

Fixed button input for controller for jump to be button A instead of B,
and thus button B for switching block. This is due to other screens make
use of A as confirm button thus users would likely get used to this
button more than others.

Level selection screen is refactored for page-based support in case we
add more levels (thus more pages) in the future.